+++ src/app/providers/products.provider.ts
@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Product, ProductAPIResponse, SortingFilter} from '../models/product';
+import { ProductRouteService } from '../services/product-route.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductsProvider {
   private readonly route: string = 'https://private-anon-4f86c5ec2a-gocco.apiary-mock.com/stores/1/products/search' ; 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient, private productRouteService:ProductRouteService) { }
 
   
   getProductsByCategory(categoryId:string):Promise<ProductAPIResponse>{
@@ -16,48 +17,20 @@ export class ProductsProvider {
   }
 
   getProductsSorted(categoryId:string, sortFilter: SortingFilter):Promise<ProductAPIResponse>{
-    let link = this.elaboratePath(categoryId,sortFilter);
+    let link = this.route + this.productRouteService.elaboratePath(categoryId,sortFilter);
       return this.http.get<ProductAPIResponse>(link).toPromise();
   }
 
   getProductsFiltered(categoryId:string, filters: any):Promise<ProductAPIResponse>{
-    let link = this.elaboratePath(categoryId, undefined,  filters);
+    let link = this.route + this.productRouteService.elaboratePath(categoryId, undefined,  filters);
     return this.http.get<ProductAPIResponse>(link).toPromise();
   }
 
   getProductsFilteredAndSorted(categoryId:string, filters:any, sortFilter: SortingFilter):Promise<ProductAPIResponse>{
-    let link = this.elaboratePath(categoryId, sortFilter, filters);
+    let link = this.route + this.productRouteService.elaboratePath(categoryId, sortFilter, filters);
     return this.http.get<ProductAPIResponse>(link).toPromise();
   }
-  elaboratePath(categoryId:string, sortFilter?:SortingFilter, filters?: any, ):string{
-    let link = this.route + "?category_id="+categoryId+'&';
-    if(sortFilter != undefined) link += "order=" +sortFilter.order + "&dir=" + sortFilter.direction +'&';
-    if(filters != undefined){
-      for(let key in filters){
-        if(key != "price" && filters[key].length!=0){
-          link += this.elaborateSubFilterPath(key, filters[key]);
-        }else if(key == "price"){
-          link += this.elaboratePricePath(key, filters[key]);
-        }
-      }
-    }  
-    let lastIndex = link.lastIndexOf('&');
-    if(lastIndex != -1 && lastIndex == link.length-1 ) link = link.substring(0,lastIndex);
-    return link;
-  }
-  elaboratePricePath(key:string, filter:any):string{
-    let filterRoute ="" ;
-    if(filter.hasOwnProperty('min')) filterRoute += `filters[price][min]=`+filter.min+'&';
-    if(filter.hasOwnProperty('max')) filterRoute+= `filters[price][max]=`+filter.max;
-    return filterRoute;
-  }
-  elaborateSubFilterPath(key:string, filter:any):string{
-    let filterRoute ="" ;
-    filter.forEach(id => {
-      filterRoute += `filters[${key}][]=`+id+'&'; 
-    });
-    return filterRoute;
-  }
+
 
   getProductByScanCode(sku:string): Promise<Product>{
     let link = "https://private-anon-3d7fb26c9f-gocco.apiary-mock.com/stores/1/products?scan_code="+sku;
